Added RTCP bridging. Separated symmitron code from sipxbridge.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12873 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXbridge/src/main/java/org/sipfoundry/sipxbridge/SipListenerImpl.java
@@ -288,12 +288,13 @@ public class SipListenerImpl implements SipListener {
                 logger.debug("dropping response " + method);
             }
 
-        } catch (Exception ex) {
+        } catch (Throwable ex) {
             Dialog dialog = responseEvent.getDialog();
-            dialog.delete();
+            if ( dialog != null ) dialog.delete();
 
             logger.error("Unexpected error processing response >>>> "
                     + response, ex);
+            logger.error("cause = " + ex.getCause());
         }
 
     }
